Improved error logging of google drive upload so that we can better debug it should an error occur.

+++ app/src/main/java/com/ndhunju/dailyjournal/util/UtilsZip.java
@@ -64,7 +64,7 @@ public class UtilsZip {
                         zout.putNextEntry(new ZipEntry(name));
                         ProgressListener.publishProgress(
                                 progressListener,
-                                (float) fileIndex / files.length,
+                                fileIndex / files.length,
                                 name
                         );
                         copy(kid, zout);
